double in A

+++ src/test/java/android/com/tronlink/wallet/mainTest/DappSendTrx.java
@@ -167,7 +167,7 @@ public class DappSendTrx extends Base {
         Double avValue =  Double.parseDouble(removeSymbolString(page.tv_balance.getText()));
         Double freezeValue =  Double.parseDouble(removeSymbolString(page.tv_freeze_amout.getText()));
         System.out.println("TotalValue:" + TotalValue);System.out.println("avValue:" + avValue);System.out.println("freezeValue:" + freezeValue);
-        Assert.assertEquals(TotalValue,avValue + freezeValue);
+        Assert.assertEquals(TotalValue.doubleValue(),avValue.doubleValue() + freezeValue.doubleValue());
     }
 
     @Test(groups = {"P0"},enabled = true, alwaysRun = true)
